Thought about looking into ImageReader with FlexRGBA_8888 format, but camera does not support that. Then foolishly tried RAW_SENSOR forgetting my MotoG4 does not support it (getting OutputSizes returns an empty array).
So  YUV_8888 -> JNI Grayscaling is all i've got. It works, but the rotation is borked and I can't for the life of me figure out why> I assume its issues with the window/texture w/h not matching image w/h or the camera output w/h and something is screwing up somewhere, but I can't get the gray area artifact to go away. Bah.

Grayscale formatting however is much higher Framerate than using Camera and doing everything in C#, so Camera2 and JNI is much more efficient so far.